update(parser): expand ProductionID to uint32 for large grammar support
- Expand ProductionID from uint16 to uint32 across language, parser, tree, and code generation
- Increase Fortran Docker limits: memory 3g→6g, LR0 core budget 160M→585M
- Add bounds checking in parse table remap to catch invalid GOTO targets
- Refactor table construction to use lazy row remapping reducing peak memory
- Update test expectations for uint32 production IDs

+++ cmd/ts2go/embedded.go
@@ -201,7 +201,7 @@ func buildParseActions(groups []ActionGroup) []gotreesitter.ParseActionEntry {
 					Symbol:            gotreesitter.Symbol(a.Symbol),
 					ChildCount:        uint8(a.ChildCount),
 					DynamicPrecedence: int16(a.Precedence),
-					ProductionID:      uint16(a.ProductionID),
+					ProductionID:      uint32(a.ProductionID),
 					Extra:             a.Extra,
 					Repetition:        a.Repetition,
 				}
